// client/src/app/api/openapi/backend/model/dtoLoadingListArticleUnit.ts
/**
 * Messe Server API
 * API für die Messe-Anwendung
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DtoLoadingListArticleUnit { 
    id?: number;
    unitId?: number;
    articleNr?: string | null;
    articleDisplayName?: string | null;
    unitWeight?: number;
    ean?: string | null;
    requiredCount?: number | null;
}